+++ src/callbacks/postPickupInit.ts
@@ -75,9 +75,7 @@ export function init(mod: Mod): void {
 export function main(pickup: EntityPickup): void {
   /*
   const entityID = getEntityID(pickup);
-  log(
-    `MC_POST_PICKUP_INIT - ${entityID} - Seed: ${pickup.InitSeed}`,
-  );
+  log(`MC_POST_PICKUP_INIT - ${entityID} - Seed: ${pickup.InitSeed}`);
   */
 
   seededGBBug.postPickupInit(pickup);