+++ src/features/optional/major/fastTravel/v.ts
@@ -108,16 +108,20 @@ function featureEnabled() {
   return config.fastTravel;
 }
 
+export function isFastTravelHappening(): boolean {
+  return v.run.state > FastTravelState.DISABLED;
+}
+
 export function setFastTravelClearFrame(): void {
   const gameFrameCount = g.g.GetFrameCount();
 
   v.room.clearFrame = gameFrameCount;
 }
 
-export function isFastTravelHappening(): boolean {
-  return v.run.state > FastTravelState.DISABLED;
-}
-
 export function setFastTravelResumeGameFrame(resumeGameFrame: int): void {
   v.level.resumeGameFrame = resumeGameFrame;
+}
+
+export function setFastTravelTookDamage(): void {
+  v.level.tookDamage = true;
 }