+++ docs/changes.md
@@ -166,7 +166,7 @@ Post-flip actions:
 
 - `0302` <!-- killExtraEnemies 45, 78 --> All extra enemies will now properly die after defeating Mom, Mom's Heart, or It Lives!
 - `0303` <!-- fastPin 62 --> Pin, Frail, Scolex, and Wormwood will now spend less time underground.
-- `0304` <!-- stopDeathSlow 66 --> Stop Death from performing his slow attack.
+- `0304` <!-- preventDeathSlow 66 --> Stop Death from performing his slow attack.
 
 > Why? It doesn't make sense that Death's slow effect can persist on the player even after Death has been defeated.
 
@@ -179,7 +179,7 @@ Post-flip actions:
 
 > Why? It can interfere with touching a trophy / chest.
 
-- `0311` <!-- stopVictoryLapPopup 273 --> The "Would you like to do a Victory Lap!?" popup will no longer appear after defeating The Lamb.
+- `0311` <!-- preventVictoryLapPopup 273 --> The "Would you like to do a Victory Lap!?" popup will no longer appear after defeating The Lamb.
 
 > Why? Speedrunners never want to do a Victory Lap; they just want to finish the run.